requirements.txt
discord.py==2.3.2
Flask==3.0.0
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.1